Move initial transition to event from model
